Created AirportController and airport dto

+++ src/main/java/com/serhatacar/flightsearchapi/repository/AirportRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AirportRepository extends JpaRepository<Airport, Integer> {
+public interface AirportRepository extends JpaRepository<Airport, Long> {
     List<Airport> findAirportsByArrivalFlightsContains(Flight flight);
     List<Airport> findAirportsByDepartureFlightsContains(Flight flight);
 }
